backend/requirements.txt
amqp==2.5.2
Babel==2.7.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
flower==0.9.3
idna==2.8
importlib-metadata==0.23
kombu==4.6.6
more-itertools==7.2.0
peewee==3.11.2
pytz==2019.3
PyYAML==5.1.2
redis==3.3.11
requests==2.22.0
tornado==5.1.1
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
